+++ bt-core/src/main/java/bt/torrent/messaging/PieceConsumer.java
@@ -120,7 +120,7 @@ public class PieceConsumer {
         connectionState.incrementDownloaded(block.length);
         if (connectionState.getCurrentAssignment().isPresent()) {
             Assignment assignment = connectionState.getCurrentAssignment().get();
-            if (pieceIndex == assignment.getPiece()) {
+            if (assignment.isAssigned(pieceIndex)) {
                 assignment.check();
             }
         }