+++ RELEASE-NOTES.md
@@ -10,6 +10,7 @@ For the latest information visit project web site: http://atomashpolskiy.github.
 
 * onTorrentStarted called twice [#117](https://github.com/atomashpolskiy/bt/issues/117)
 * Add new onMetadataAvailable event (fired after torrent's metadata has been fetched)
+* Continuous piece assignments (better request pipelining)
 
 ## 1.8